bot/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-multiselectfield==0.1.12
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
phonenumbers==8.12.35
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-multiselectfield==0.1.12
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
phonenumbers==8.12.35
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-multiselectfield==0.1.12
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
phonenumbers==8.12.35
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-multiselectfield==0.1.12
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
phonenumbers==8.12.35
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-multiselectfield==0.1.12
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
phonenumbers==8.12.35
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
phonenumbers==8.12.35
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
